Update create and new action
Previously we were not checking what happens when @game.save is false
Now when the form doesn't save means there is some invalid input then it
will reload the form.
We use render instead of redirect_to as render will send the instance of
game to the new template when it's rendered while redirect_to will
refresh the page again.

New action will create a new instance of the game.