Render partial collections and partial form to clean up the code
Create _guess.html.erb. The partial in the file will be rendered
once for each guess that is in @game.guesses collection.
The render method will iterate over @game.guesses collection, it assigns
each guess to a local variable named the same as the partial which is guess

Create _form.html.erb under guesses view which will define partial template
we want to render guesses/form